backend/pyproject.toml
[tool.poetry]
authors = ["Oleksandr Korol <zibertua@gmail.com>"]
description = ""
name = "backend"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
Django = "^4.2.4"
django-cors-headers = "^4.2.0"
django-extensions = "^3.2.3"
django-filter = "^23.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.0"}
drf-spectacular = "^0.26.4"
gunicorn = "^21.2.0"
python = "^3.11"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
psycopg = {extras = ["binary"], version = "^3.1.10"}
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.1.0"}
pillow = "^10.0.1"

[tool.poetry.group.dev.dependencies]
mixer = "^7.2.2"
ruff = "^0.1.3"

[tool.poetry.group.test.dependencies]
coverage = {extras = ["toml"], version = "^7.2.7"}
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.ruff]
fixable = ["ALL"]
ignore = ["RUF005"]
line-length = 120
select = ["E", "F", "W", "I", "I001", "C4", "Q", "RUF", "PIE", "COM"]
# Exclude a variety of commonly ignored directories.
exclude = [
  ".git",
  ".mypy_cache",
  ".ruff_cache",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "migrations",
  "dist",
  ".venv",
  ".conf",
  "asgi.py",
]
per-file-ignores = {}
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py311"
[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10
[tool.black]
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | ^.*\b(migrations)\b.*$
)/
'''
line-length = 120
target-version = ["py311"]

[tool.coverage.run]
branch = true
omit = [
  ".venv/*",
  "*/migrations/*",
  "*/admin.py",
  "*/apps.py",
  "manage.py",
  "*/settings.py",
  "*/asgi.py",
  "*/wsgi.py",
  "*/tests.py",
  "example/*",
  "shop/*",
]

[tool.coverage.report]
show_missing = true
skip_empty = true
sort = "cover"
